+++ test/common/__init__.py
@@ -3,5 +3,5 @@ from .infrastructure import *  # noqa: F401
 from .sugar import *  # noqa: F401
 from .testbenchio import *  # noqa: F401
 from .profiler import *  # noqa: F401
-from .coreblocks import *  # noqa: F401
+from .assertion import *  # noqa: F401
 from transactron.utils import data_layout  # noqa: F401
