+++ test/regression/pysim.py
@@ -1,6 +1,8 @@
 from amaranth.sim import Passive, Settle
 from amaranth.utils import log2_int
 
+from transactron.utils.dependencies import DependencyManager
+
 from .memory import *
 from .common import SimulationBackend
 
@@ -15,7 +17,7 @@ from coreblocks.peripherals.wishbone import WishboneBus
 
 class PySimulation(SimulationBackend):
     def __init__(self, verbose: bool, traces_file: Optional[str] = None):
-        self.gp = GenParams(full_core_config)
+        self.gen_params = GenParams(full_core_config)
         self.running = False
         self.cycle_cnt = 0
         self.verbose = verbose
@@ -30,7 +32,7 @@ class PySimulation(SimulationBackend):
             while True:
                 yield from wb_ctrl.slave_wait()
 
-                word_width_bytes = self.gp.isa.xlen // 8
+                word_width_bytes = self.gen_params.isa.xlen // 8
 
                 # Wishbone is addressing words, so we need to shift it a bit to get the real address.
                 addr = (yield wb_ctrl.wb.adr) << log2_int(word_width_bytes)
@@ -90,9 +92,9 @@ class PySimulation(SimulationBackend):
         return f
 
     async def run(self, mem_model: CoreMemoryModel, timeout_cycles: int = 5000) -> bool:
-        wb_instr_bus = WishboneBus(self.gp.wb_params)
-        wb_data_bus = WishboneBus(self.gp.wb_params)
-        core = Core(gen_params=self.gp, wb_instr_bus=wb_instr_bus, wb_data_bus=wb_data_bus)
+        wb_instr_bus = WishboneBus(self.gen_params.wb_params)
+        wb_data_bus = WishboneBus(self.gen_params.wb_params)
+        core = Core(gen_params=self.gen_params, wb_instr_bus=wb_instr_bus, wb_data_bus=wb_data_bus)
 
         m = SimpleTestCircuit(core)
 
@@ -105,11 +107,12 @@ class PySimulation(SimulationBackend):
         def my_assert(val, msg):
             assert val, msg
 
-        sim = PysimSimulator(m, max_cycles=timeout_cycles, traces_file=self.traces_file)
+        clk_period = 1e-6
+        sim = PysimSimulator(m, max_cycles=timeout_cycles, traces_file=self.traces_file, clk_period=clk_period)
         sim.add_sync_process(self._wishbone_slave(mem_model, wb_instr_ctrl, is_instr_bus=True))
         sim.add_sync_process(self._wishbone_slave(mem_model, wb_data_ctrl, is_instr_bus=False))
         sim.add_sync_process(self._waiter())
-        sim.add_sync_process(make_assert_handler(self.gp, my_assert))
+        sim.add_sync_process(make_assert_handler(self.gen_params.get(DependencyManager), my_assert, clk_period))
         res = sim.run()
 
         if self.verbose:
